tests(array): split the pop/push tests into a sibling file for the 2000-line gate
